requirements.txt
ml-collections==0.1.0
tensorflow-gan==2.0.0
tensorflow_io
tensorflow_datasets==3.1.0
tensorflow==2.4.0
tensorflow-addons==0.12.0
tensorboard==2.4.0
tensorflow-probability==0.12
absl-py==0.10.0
torch>=1.7.0
torchvision
ninja
scipy
numpy
matplotlib
seaborn
pandas
jax
jaxlib
sklearn
celluloid
tensorflow
prdc